:truck: move sections into component
